Simplify HTML attribute token list merging
Rely on view context's `#token_list` to merge token lists.

Manipulating the token list manually (through calls to `String#split`
and `Array#join`) had the potential to affect
`ActiveSupport::SafeBuffer` HTML safety.

This commit replaces that bespoke logic with calls to
[ActionView::Helpers::TagHelper#token_list][] to rely entirely on the
view for managing the escaping, safety, and joining of the String
values.

[]: https://edgeapi.rubyonrails.org/classes/ActionView/Helpers/TagHelper.html#method-i-token_list